# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.App:
    properties:
      createdAt:
        type: string
      enableAlarm:
        type: boolean
      id:
        type: integer
      kafkaInputTopic:
        type: string
      kafkaOutputTopic:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.LogTemplate:
    properties:
      app:
        type: string
      clusterId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      level:
        type: string
      size:
        type: integer
      tokens:
        type: string
      updatedAt:
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.LogTemplateSearch:
    properties:
      app:
        type: string
      clusterId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      level:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      size:
        type: integer
      tokens:
        type: string
      updatedAt:
        type: string
    type: object
...
  request.PageInfo:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /app/app:
    delete:
      consumes:
      - application/json
      parameters:
      - description: AppID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.App'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除App
      tags:
      - App
    post:
      consumes:
      - application/json
      parameters:
      - description: App名, kafka输入topic, kafka输出topic
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.App'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建App
      tags:
      - App
    put:
      consumes:
      - application/json
      parameters:
      - description: AppID, 启用报警
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.App'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新app, 是否启用报警
      tags:
      - App
  /app/appList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取权限App列表
      tags:
      - App
...
  /logTemplate/createLogTemplate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建LogTemplate
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.LogTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建LogTemplate
      tags:
      - LogTemplate
  /logTemplate/deleteLogTemplate:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除LogTemplate
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.LogTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除LogTemplate
      tags:
      - LogTemplate
  /logTemplate/deleteLogTemplateByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除LogTemplate
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除LogTemplate
      tags:
      - LogTemplate
  /logTemplate/getLogTemplateList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取LogTemplate列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.LogTemplateSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取LogTemplate列表
      tags:
      - LogTemplate
  /logTemplate/getRealtimeResult:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取实时的日志模版
      tags:
      - LogTemplate
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"